I've finished a basic implementation of the new design.
Note: This particular commit is not intended to be used as a valid version of the project.  It won't even compile if you try.

I'm making this commit because while working on this redesign I tried quite a few different ideas, and there are a bunch of experiments and alternate versions of things, that I want to save so that I can revisit them later if I want.

I'm also about to make another set of changes to consolidate a lot of the overlap between the Factor classes and the CausalFactor classes, and I want a place I can come back to if I mess everything up somehow.