I'm going to save this here, before I attempt to change it so that Factors handle their own dependents, instead of letting the cores manage them.  This is mostly because I want to make it so that you can swap out a Reactor's core at will, allowing it to change its behavior on the fly.  Factors are going to expect that their subscriptions are still valid though, so the subscriptions need to be moved out of the cores.
Renamed Priority to Update Priority, because I plan to create other classes that have other unrelated priorities.

Renamed the CollectionStates to CollectionCores.

Moved IDeterminant into its own file.

Changed IState<T> so that it is now used for Factors you can set the value of.  If you can only get the value it should be an IFactor<T>.

Created IReactor<T>, which as you might expect is for use when  you want people to be able to access the methods that control a Reactor in addition to being able to access its value.

 Changed IObservedFactor<T> to inherit IFactor<T>.

 Removed the Peek() method from the Proactive<T> and Reactive<T> classes because it doesn't really fit there anymore, since only observed factors use it.

 Moved a bunch of the tests dealing with functions that pertain to the IDeterminant interface to their own file.

 Fixed the ConstructorTester classes.  They had the wrong arguments listed in their TestFixture attribute.

 Cleaned up and renamed some test related things.