Created an alternate version of the Factor and Reactor classes where the Factors store their own dependents, as opposed to them having their core handle them.  The reason for the change is that I want to try to make it so that Factor cores can be swapped out on the fly, allowing you to change their behavior without their dependents having to unsubscribe and resubscribe to some new Factor.
I have both versions of the Factors in this commit in case I want to come back to either one later.  The newer versions are creatively named Factor2, Reactor2, etc.

Changed the NumberOfTimesReacted field for IReactors to VersionNumber and moved it to IFactor.  This was mostly so that Factor subscribers could store the version number of a trigger, so that later they could be certain it hadn't changed.

Created a Settings class and an interface for handling logging errors in Factors.  I moved some of the error messages to an implementation of the logging interface.