Changed the HasReacted property in ReactorCore so that it's a bool that's set during GenerateOutcome(), to avoid situations where the property incorrectly says the core hasn't reacted because the first time a reaction happened nothing changed and the version number didnt increment.  (An example of this would be a Reactive<T> whose initial reaction resulted in T's default value.  Since its uninitialized value before the reaction would have already been the default value, nothing will be considered to have changed, and the version number won't increment.)
Added parameters to the CoreTriggered() and CoreDestabilized() methods.  Right now they're mostly used to check that the core calling the methods is the current core and not a previous one that was swapped out but is still somehow linked to the reactor.

Simplified Destabilize() because it seemed like it was checking conditions that were no longer relevant.