Did a large rework of how the triggering process works.  I had the idea that I could design the system so that a single factor could trigger its subscribers in a variety of different ways by communicating using a series of flags.  The class triggering the subscribers can come up with a collection of flags representing the different events that might trigger a subscriber to need to update.  When the triggering class decides to use the Observer class in one of its methods to notify it's involved with the outcome of a process, it passes a set of the trigger event flags that includes only the events that would invalidate whatever action/result they were involved in. As an example, if a process uses a Reactive<double> to test if X > value of the Reactive, and it's true, if the Reactive's value is change to a lower value later, the result of that earlier test is still valid so there's no need to trigger that particular computation to update (as opposed to the current system where all value changes trigger all subscribers).
Reworked the way Modifiables are structured so that applying modifiers is much simpler.

Created a bunch factors, useful for representing and interacting with times and dates. Still pretty unfinished.

Added the Condition class which is a Reactive<bool> but with additional methods that make it easier to write simple readable code

Added the ICondition interface to match the Condition class.

Added the IEnumerableFactor<T> interface which is what it sounds like, a Factor that can be enumerated.

Added the Peek() method back to the value Factors as a way to get a Factor's current value without triggering it to update.  Useful for Reactions that are triggered when a another Factor's value changes, and need to get that value to execute the Reaction, but don't want to trigger the Factor to update again when it retrieves its value.

Added a NumberOfNecessarySubscribers property to the Influence class and IInfluence.

Refactored/added a lot more than is just listed here.

As a note, there is going to be a bunch of unused classes, interfaces, etc in this version.  That's partially to save some of it in case I want to revisit it later, and partially because I need to go through and get rid of the things I experimented with when designing the changes, but am not going to use.