Added and changed a large amount of code related to collection Factors.  I'm currently to building a new set of these collections that are intended to merge a bunch of the disparate collection Factors classes that overlap each other in terms of functionality and purpose, as well as add some new features.
Moved all of the methods from the ICollectionCore<T> class to IProactiveCollectionCore<T>, since most of them were things a Reactive collection wouldn't offer like adding items.

The ICollectionCore<T> interface now contains a single method CollectionEquals(), which provides a way for any collection core to compare its collection to one owned by a similar core.  This was important for making it so that collection cores could compare themselves when a Factor swaps cores.

Created the 'Implementor' classes which are intended to be embedded in the new collection Factors I'm working on.  There is one of these classes for each type of collection (List, HashSet, Dictionary).  The class is designed to hold the majority of the implementation for making that type of collection work as a Factor, which should allow me to use the same class to enable making multiple types of Factors based on a particular collection type, without duplicating the code across all of them.  For example using a ListImplementor to perform all of the common functions shared across a ReactiveList<T>, a ProactiveList<T>, and any other List based collection Factor.